==> Find_Company_Permnos/01_Create_First_Patent_Assignment.ipynb <==
import xml.etree.ElementTree as ET 
import pandas as pd
import codecs
import os
import zipfile
import math
import numpy as np
import glob
from joblib import Parallel, delayed

os.chdir("C:\\Users\\fjying\\Desktop\\RA_Columbia\\patentparsing\\identifycompanypermno\\input")

files_list = os.listdir("C:\\Users\\fjying\\Desktop\\RA_Columbia\\patentparsing\\identifycompanypermno\\input")

files_list[0:-2]

total_patents = pd.read_csv("2013_01_notmeaningfultransfer.csv")

total_patents  = total_patents.sort_values(by =['recorded_date', 'last_update_date'])
total_patents  = total_patents.drop_duplicates( ['recorded_date', 'patent_ids'], keep = 'first')
total_patents['patent_ids'] = total_patents['patent_ids'].astype(str)

total_patents = total_patents[['patent_ids', 'assignor', 'name_0', 'recorded_date', 'last_update_date']]

files_all = files_list[1:-2]
for file in files_all:
    patents = pd.read_csv(file)
    patents['patent_ids'] = patents['patent_ids'].astype(str)
    patents = patents[['patent_ids', 'assignor', 'name_0', 'recorded_date', 'last_update_date']]
    total_patents= pd.concat([total_patents, patents])
    total_patents  = total_patents.sort_values(by =['recorded_date', 'last_update_date'])
    total_patents  = total_patents.drop_duplicates( ['recorded_date', 'patent_ids'], keep = 'first')

total_patents = total_patents.sort_values('recorded_date')

total_patents.to_csv("C:\\Users\\fjying\\Desktop\\RA_Columbia\\patentparsing\\identifycompanypermno\\intermediate_output\\first_patent_assignment_1980_2015.csv", index = False)
