==> Find_Company_Permnos/03_Build_FirstAssignorName_Permno_Crosswalk.ipynb <==
import pandas as pd
import datetime
import numpy as np
import os
from string_grouper import match_strings

rawpatents = pd.read_csv("C:\\Users\\fjying\\Desktop\\RA_Columbia\\patentparsing\\identifycompanypermno\\input\\rawpatentrecord.csv")

os.chdir("C:\\Users\\fjying\\Desktop\\RA_Columbia\\patentparsing\\identifycompanypermno\\intermediate_output")

firstpatentassignments = pd.read_csv("first_patent_assignment_1980_2015_standardized.csv")

firstpatentassignments = firstpatentassignments[~firstpatentassignments['rawname_0'].isna()]

rawpatents = rawpatents.dropna(subset = ['patnum', 'permno'])
rawpatents = rawpatents.drop_duplicates(['patnum', 'permno'])
rawpatents = rawpatents.rename(columns = {'patnum': 'patent_ids'})

rawpatents['patent_ids'] = rawpatents['patent_ids'].astype(str)
firstpatentassignments['patent_ids'] = firstpatentassignments['patent_ids'].astype(str)

duplicatedpatentids = np.unique(rawpatents['patent_ids'][rawpatents['patent_ids'].duplicated()].values)

rawpatents_noduplicated = rawpatents[~rawpatents['patent_ids'].isin(duplicatedpatentids)]
rawpatents_duplicated = rawpatents[rawpatents['patent_ids'].isin(duplicatedpatentids)]

crosswalk = pd.merge(rawpatents_noduplicated, firstpatentassignments, how= 'inner')

crosswalk = crosswalk[['name_0', 'permno', 'recorded_date', 'patent_ids']]

crosswalk = crosswalk.sort_values(['recorded_date'])

crosswalk = crosswalk.drop_duplicates(['name_0', 'permno'])

#Resolve relationship between one permno and multiple company names
duplicated_permnos = np.unique(crosswalk['permno'][crosswalk['permno'].duplicated()].values)

crosswalk_one_on = crosswalk[~crosswalk['permno'].isin(duplicated_permnos)]
crosswalk_multiple_on_one = crosswalk[crosswalk['permno'].isin(duplicated_permnos)]

crosswalk_final = pd.DataFrame()
for i in duplicated_permnos:
    temp = crosswalk[crosswalk['permno'] == i]
    temp = temp.drop_duplicates(subset = ['name_0'], keep = 'first')
    try:
        matches = match_strings(temp['name_0'])
    except ValueError:
        print("only stop words")
        continue
    matches = matches[matches['left_side'] != matches['right_side']]
    if matches.empty:
        continue
    name = matches['left_side'].value_counts().keys()[0]
    crosswalk_final = crosswalk_final.append(temp[temp['name_0'] == name])

crosswalk_final = crosswalk_final.append(crosswalk_one_on)

#resolve one company name multiple permnos due to incomplete patent records
crosswalk_final_one_comp_multiple_permno = crosswalk_final[crosswalk_final['name_0'].duplicated(keep = False)].sort_values(['name_0'])

crosswalk_final_one_on_one = crosswalk_final[~crosswalk_final['name_0'].duplicated(keep = False)]

crosswalk_final_one_comp_multiple_permno.to_csv("crosswalk_final_one_comp_multiple_permno.csv", index = False)

crsp_directory = pd.read_csv("C:\\Users\\fjying\\Desktop\\RA_Columbia\\patentparsing\\identifycompanypermno\\input\\crsp_directory.csv")

for i in crosswalk_final_one_comp_multiple_permno['permno'].values:
    print(i)
    display(crsp_directory[crsp_directory['PERMNO'] == i]['NAME OF FIRM'].values)

temp = pd.read_csv("crosswalk_final_one_comp_multiple_permno_resolved.csv")
crosswalk_final = crosswalk_final_one_on_one.append(temp)

crosswalk_final = crosswalk_final.drop_duplicates(['name_0', 'permno'])

#link first patent assignments with raw patent records with duplicated patent ids

crosswalk_2 = pd.merge(rawpatents_duplicated, firstpatentassignments, how= 'inner')

crosswalk_final['name_0'] = crosswalk_final['name_0'].astype(str)
crosswalk_2['name_0'] = crosswalk_2['name_0'].astype(str)

crosswalk_2 = crosswalk_2[~crosswalk_2['name_0'].isin(np.unique(crosswalk_final['name_0'].values))]

crosswalk_2 = crosswalk_2[~crosswalk_2['permno'].isin(np.unique(crosswalk_final['permno'].values))]

crosswalk_2 = crosswalk_2[['name_0', 'permno', 'recorded_date', 'patent_ids']]

crosswalk_2 = crosswalk_2.drop_duplicates(['name_0', 'permno'])

#resolve one company name but multiple permnos
#drop all entries
crosswalk_2 = crosswalk_2.drop_duplicates(['name_0'], keep = False)
duplicated_companynames = np.unique(crosswalk_2['name_0'][crosswalk_2['name_0'].duplicated()].values)

#resolve one permno but multiple company names
duplicated_permnos = np.unique(crosswalk_2['permno'][crosswalk_2['permno'].duplicated()].values)
crosswalk_2_one_on = crosswalk_2[~crosswalk_2['permno'].isin(duplicated_permnos)]
crosswalk_2_multiple_on_one =crosswalk_2[crosswalk_2['permno'].isin(duplicated_permnos)]

#resolve one company name but multiple permnos
#drop the cases due to multiple patent ids 
crosswalk_final_2 = pd.DataFrame()
for i in duplicated_permnos:
    temp = crosswalk_2_multiple_on_one[crosswalk_2_multiple_on_one['permno'] == i]
    temp = temp.drop_duplicates(subset = ['name_0'], keep = 'first')
    try:
        matches = match_strings(temp['name_0'])
    except ValueError:
        print("only stop words")
        continue
    matches = matches[matches['left_side'] != matches['right_side']]
    if matches.empty:
        continue
    name = matches['left_side'].value_counts().keys()[0]
    crosswalk_final_2 = crosswalk_final_2.append(temp[temp['name_0'] == name])

crosswalk_2 = crosswalk_2_one_on.append(crosswalk_final_2)

#Check final crosswalk
crosswalk_final = pd.concat([crosswalk_final, crosswalk_2])
crosswalk_final[crosswalk_final['permno'].duplicated(keep = False)].sort_values(['permno'])

#old records are not kept
#check these companies indeed have multipe permnos
crosswalk_final[crosswalk_final['name_0'].duplicated(keep = False)].sort_values(['name_0'])


crosswalk_final.loc[21, 'name_0'] = 'heico'

crosswalk_final = crosswalk_final.drop([30930])
crosswalk_final = crosswalk_final.drop([931871])
crosswalk_final = crosswalk_final.drop([95826])
crosswalk_final = crosswalk_final.drop([30])
crosswalk_final = crosswalk_final.drop([89259])

crosswalk_final.to_csv("permno_processedfirstassignorname_crosswalk.csv", index = False)
