==> Find_Company_Permnos/05_Find_Permnos.ipynb <==
import pandas as pd
from joblib import Parallel, delayed
import os

os.chdir("C:\\Users\\fjying\\Desktop\\RA_Columbia\\notmeaningfultransfer")

files = os.listdir("C:\\Users\\fjying\\Desktop\\RA_Columbia\\notmeaningfultransfer")

def find_permnos(file, crosswalk):
    patents  = pd.read_csv(file)
    columns_new = []
    columns = patents.columns
    
    intersected_company_names = set(patents['assignor'].dropna().values).intersection(set(crosswalk['company'].values))
    temp = crosswalk[crosswalk['company'].isin(intersected_company_names)]
    company_permno_dict = dict(zip(temp['company'].values, temp['permno'].values))
    patents['sellerpermno'] = patents['assignor'].map(company_permno_dict)
    columns_new.extend(['assignor', 'sellerpermno'])
    
    i = 0
    while 'name_' + str(i)  in columns:
        buyer = 'name_' + str(i)
        intersected_company_names = set(patents[buyer].dropna().values).intersection(set(crosswalk['company'].values))
        temp = crosswalk[crosswalk['company'].isin(intersected_company_names)]
        company_permno_dict = dict(zip(temp['company'].values, temp['permno'].values))
        patents['buyerpermno_' + str(i)] = patents[buyer].map(company_permno_dict)
        columns_new.extend([buyer, 'buyerpermno_' + str(i)])
        i = i + 1    
    columns_list = ['notmeaningful', 'patent_ids', 'recorded_date', 'reel-no', 'frame-no', 'total_number_assignees']
    columns_list.extend(columns_new)
    columns_list.extend(list(set(columns)-set(columns_list)))
    patents = patents[columns_list]
    patents.to_csv(file[:-4] +"_final.csv", index = False)
    #drop empty columns which we cannot find permnos for multiple assignees
    return patents

crosswalk = pd.read_csv("C:\\Users\\fjying\\Desktop\\RA_Columbia\\patentparsing\\identifycompanypermno\\intermediate_output\\permno_processedallcompanyames_crosswalk.csv")

files

#notmeanigful patent transfers files
#If the transfer is not meaningful, same permno

#multiprocessing for files
with Parallel(n_jobs=-1, verbose=1, prefer="threads") as parallel:
    parallel(delayed(find_permnos)(file, crosswalk) for file in files)
